Improvements in map making: keep image information such as distance, psf_filter, etc. in map FITS files, better checking of present maps

+++ core/tools/filesystem.py
@@ -591,6 +591,10 @@ def remove_directory(path):
     :return:
     """
 
+    # Test
+    if not is_directory(path): raise ValueError("Is not a directory: '" + path + "'")
+
+    # Remove
     shutil.rmtree(path)
 
 # -----------------------------------------------------------------
